Separated IstioGatewayPolicyGenerator into a separate file.
